+++ airframe-http-codegen/src/test/scala/example/rpc/RPCTestService.scala
@@ -12,8 +12,7 @@
  * limitations under the License.
  */
 package example.rpc
-import wvlet.airframe.http.RPC
-import wvlet.airframe.http.router.{RxRouter, RxRouterProvider}
+import wvlet.airframe.http.{RPC, RxRouter, RxRouterProvider}
 
 /**
   */
