+++ airframe-http-codegen/src/test/scala/wvlet/airframe/http/codegen/RPCClientGeneratorTest.scala
@@ -13,7 +13,7 @@
  */
 package wvlet.airframe.http.codegen
 
-import wvlet.airframe.http.router.RxRouter
+import wvlet.airframe.http.RxRouter
 import wvlet.airspec.AirSpec
 
 class RPCClientGeneratorTest extends AirSpec {
