+++ airframe-http/src/main/scala/wvlet/airframe/http/RxRouterProvider.scala
@@ -11,16 +11,23 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package wvlet.airframe.http.router
+package wvlet.airframe.http
 
-import wvlet.airframe.http.RxHttpFilter
-import scala.language.experimental.macros
-
-trait RxRouterObjectBase {
-  def of[Controller]: RxRouter = macro RxRouterMacros.of[Controller]
-  def filter[Filter <: RxHttpFilter]: RxRouter.FilterNode = macro RxRouterMacros.filter[Filter]
-}
-
-trait RxRouteFilterBase {
-  def andThen[Filter <: RxHttpFilter]: RxRouter.FilterNode = macro RxRouterMacros.andThenFilter[Filter]
+/**
+  * An interface used for RPC clients (sbt-airframe) to discover the default router for the RPC endpoint.
+  *
+  * Example usage:
+  * {{{
+  *   trait MyRPC {
+  *     def hello(name:String) : String
+  *   }
+  *
+  *   object MyRPC extends RxRouterProvider {
+  *     // sbt-airframe will generate an RPC client using this router
+  *     override def router: RxRouter = RxRouter.of[MyRPC]
+  *   }
+  * }}}
+  */
+trait RxRouterProvider {
+  def router: RxRouter
 }
